requirements.txt
aiogram==3.27.0
anthropic==0.96.0
gspread==6.2.1
google-auth==2.49.2
pydantic-settings==2.13.1
python-dotenv==1.2.2
